+++ android-twa/app/src/main/java/com/taatvam/rti/DelegationService.java
@@ -1,0 +1,14 @@
+package com.taatvam.rti;
+
+
+
+public class DelegationService extends
+        com.google.androidbrowserhelper.trusted.DelegationService {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+        
+    }
+}
+